aep-prototypes/brandscrape-studio/requirements.txt
flask>=3.1
requests>=2.32
beautifulsoup4>=4.13
lxml>=5.3
playwright>=1.50
Pillow>=11.1
cssutils>=2.11
openai>=1.68
anthropic>=0.49
python-dotenv>=1.1
